remove all edit mode logic
