+++ apps/chat/src/components/Chat/Publish/PublicationHandler/ReviewRowItems/PublicationFolderRow.tsx
@@ -20,9 +20,6 @@ interface Props<T extends PublicationReviewItem> {
     item: T;
     level: number;
   }>;
-
-  isEditable: boolean;
-  editedName: string;
 }
 
 const filteredItems = <T extends PublicationReviewItem | FolderInterface>(
@@ -33,14 +30,14 @@ const filteredItems = <T extends PublicationReviewItem | FolderInterface>(
     allItems.filter((item) => item.folderId === currentFolderId) as T[],
   );
 
+const isEditMode = false;
+
 export const PublicationFolderRow = <T extends PublicationReviewItem>({
   currentFolder,
   allFolders,
   allItems,
   level,
   itemComponent,
-  isEditable,
-  editedName,
 }: Props<T>) => {
   const [isFocused, setIsFocused] = useState(false);
 
@@ -72,13 +69,13 @@ export const PublicationFolderRow = <T extends PublicationReviewItem>({
             className="relative flex-1 select-none truncate text-left"
             data-qa="folder-name"
           >
-            {isEditable ? (
+            {isEditMode ? (
               <div className="block flex-1 truncate whitespace-pre break-all text-left text-primary">
                 <input
                   onFocus={() => setIsFocused(true)}
                   onBlur={() => setIsFocused(false)}
                   className="h-[24px] w-full border-b border-primary bg-layer-2 px-1 py-[2px] text-sm text-primary placeholder:text-secondary focus:border-accent-primary focus:outline-none"
-                  value={editedName}
+                  value={currentFolder.name}
                   // eslint-disable-next-line @typescript-eslint/no-unused-vars, @typescript-eslint/no-empty-function
                   onChange={(e) => {}}
                 />
@@ -106,8 +103,6 @@ export const PublicationFolderRow = <T extends PublicationReviewItem>({
               itemComponent={itemComponent}
               allItems={allItems}
               allFolders={allFolders}
-              isEditable={isEditable}
-              editedName={editedName}
             />
           ))}
         </div>
