+++ apps/chat/src/components/Chat/Publish/PublicationHandler/ReviewRowItems/PublicationPromptRow.tsx
@@ -7,25 +7,18 @@ import { PromptInfo } from '@epam/ai-dial-shared';
 interface Props {
   prompt: PromptInfo;
   level: number;
-  isEditable: boolean;
-  editedName: string;
 }
 
-export const PublicationPromptRow: React.FC<Props> = ({
-  prompt,
-  level,
-  isEditable,
-  editedName,
-}) => {
+export const PublicationPromptRow: React.FC<Props> = ({ prompt, level }) => {
   return (
     <PublicationItemRow
       level={level}
-      isEditable={isEditable}
-      editedName={editedName}
       name={prompt.name}
       Icon={<IconBulb size={18} className="text-secondary" />}
       publicationInfo={prompt.publicationInfo}
       dataQa="prompt"
+      editedName={prompt.name}
+      isEditMode={false}
     />
   );
 };
